# backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.BatchStatus:
    enum:
    - scheduled
    - active
    - frozen
    - finished
    type: string
    x-enum-varnames:
    - StatusScheduled
    - StatusActive
    - StatusFrozen
    - StatusFinished
  models.BatchType:
    enum:
    - REGULAR
    - MAKEUP
    type: string
    x-enum-varnames:
    - BatchRegular
    - BatchMakeup
  models.ExamBatch:
    properties:
      allowedParticipants:
        description: Stored as JSON
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      duration:
        description: minutes
        type: integer
      endTime:
        type: string
      frozenAt:
        type: string
      id:
        type: string
      quizId:
        type: string
      resumedAt:
        type: string
      startTime:
        type: string
      status:
        $ref: '#/definitions/models.BatchStatus'
      title:
        type: string
      type:
        $ref: '#/definitions/models.BatchType'
      waitlist:
        description: Stored as JSON
        type: string
    type: object
  models.ExamType:
    enum:
    - daily_quiz
    - midterm
    - final
    - practice
    type: string
    x-enum-varnames:
    - ExamDaily
    - ExamMidterm
    - ExamFinal
    - ExamPractice
  models.Question:
    properties:
      correctAnswer:
        description: For non-MCQ
        type: string
      explanation:
        type: string
      id:
        type: string
      options:
        items:
          $ref: '#/definitions/models.QuestionOption'
        type: array
      orderIndex:
        type: integer
      points:
        type: integer
      quizId:
        type: string
      text:
        type: string
      type:
        $ref: '#/definitions/models.QuestionType'
    type: object
  models.QuestionOption:
    properties:
      id:
        type: string
      isCorrect:
        type: boolean
      questionId:
        type: string
      text:
        type: string
    type: object
  models.QuestionType:
    enum:
    - mcq
    - true_false
    - short_answer
    - essay
    type: string
    x-enum-varnames:
    - TypeMCQ
    - TypeTrueFalse
    - TypeShortAnswer
    - TypeEssay
  models.Quiz:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        type: string
      examType:
        $ref: '#/definitions/models.ExamType'
      id:
        type: string
      passingScore:
        type: integer
      questions:
        items:
          $ref: '#/definitions/models.Question'
        type: array
      subjectId:
        type: string
      title:
        type: string
      totalPoints:
        type: integer
      updatedAt:
        type: string
    type: object
  models.User:
    properties:
      avatarUrl:
        type: string
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      institutionId:
        type: string
      name:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
    type: object
  models.UserRole:
    enum:
    - admin
    - teacher
    - student
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleTeacher
    - RoleStudent
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
  description: This is the API server for Academic Suite.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Academic Suite API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT tokens
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User Login
      tags:
      - auth
  /api/auth/profile:
    get:
      consumes:
      - application/json
      description: Get the profile of the currently logged in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get User Profile
      tags:
      - auth
  /api/batches:
    get:
      description: Retrieve a list of all exam batches
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ExamBatch'
            type: array
      summary: Get All Exam Batches
      tags:
      - batches
    post:
      consumes:
      - application/json
      description: Create a new exam batch
      parameters:
      - description: Batch Data
        in: body
        name: batch
        required: true
        schema:
          $ref: '#/definitions/models.ExamBatch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExamBatch'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create Exam Batch
      tags:
      - batches
  /api/batches/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of an exam batch
      parameters:
      - description: Batch ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Status Object (e.g. {''status'': ''active''})'
        in: body
        name: status
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExamBatch'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update Batch Status
      tags:
      - batches
  /api/quizzes:
    get:
      description: Retrieve a list of all quizzes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Quiz'
            type: array
      summary: Get All Quizzes
      tags:
      - quizzes
    post:
      consumes:
      - application/json
      description: Create a new quiz with questions
      parameters:
      - description: Quiz Data
        in: body
        name: quiz
        required: true
        schema:
          $ref: '#/definitions/models.Quiz'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Quiz'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create New Quiz
      tags:
      - quizzes
  /api/quizzes/{id}:
    get:
      description: Retrieve a single quiz by its ID
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Quiz'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Quiz by ID
      tags:
      - quizzes
swagger: "2.0"